Ai/requirements.txt
Flask==2.0.1           
Werkzeug==2.2.3 
pytesseract==0.3.10
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
# Face-match dependencies:
deepface==0.0.79
tensorflow==2.11.0      
keras==2.11.0    